--- zingo_wallet/src/version.rs
@@ -1,1 +1,0 @@
-pub struct Version
